Arreglos tipos de dato de TareaConfiguracion

+++ APPWEB-INCOMEMotorCalculoJava-model/src/test/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaConfiguracionRepositoryCustomImplTest.java
@@ -54,7 +54,7 @@ public class TareaConfiguracionRepositoryCustomImplTest {
         TareaConfiguracion entity = mock(TareaConfiguracion.class);
         when(entity.getTarea()).thenReturn(tarea);
         when(entity.getCclIdOrigen()).thenReturn("ID ORIGEN");
-        when(entity.getIcmIdConfig()).thenReturn(87654L);
+        when(entity.getIcmIdConfig()).thenReturn(87654);
         when(entity.getFechaInicio()).thenReturn(TimeUtils.toDate(LocalDate.of(2015, 1, 1)));
         when(entity.getFechaFin()).thenReturn(TimeUtils.toDate(LocalDate.of(2015, 1, 31)));
         when(entity.getIcmIdTpHora()).thenReturn("ID TP HORA");
@@ -66,7 +66,7 @@ public class TareaConfiguracionRepositoryCustomImplTest {
         // ICM_ID_TP_HORA, ICM_CK_VENTA_IMPUESTOS
         verify(pstmt, times(1)).setLong(1, tarea.getId());
         verify(pstmt, times(1)).setString(2, entity.getCclIdOrigen());
-        verify(pstmt, times(1)).setLong(3, entity.getIcmIdConfig());
+        verify(pstmt, times(1)).setInt(3, entity.getIcmIdConfig());
         verify(pstmt, times(1)).setObject(4, entity.getFechaInicio());
         verify(pstmt, times(1)).setObject(5, entity.getFechaFin());
         verify(pstmt, times(1)).setString(6, entity.getIcmIdTpHora());
